Backup train_gpt2.py
The current train_gpt2.py is closing to the final version, but it's a
bit ahead of its time. In order to allow each step in the video to be
hands-on, I temporarily back it up. A new train_gpt2.py will be created
and modified step by step, rather than doing everything in one go.